handle missleading warning

+++ apps/pwabuilder/Services/IOS/TempDirectory.cs
@@ -52,7 +52,10 @@
                 {
                     try
                     {
-                        Directory.Delete(directory, recursive: true);
+                        if (Directory.Exists(directory))
+                        {
+                            Directory.Delete(directory, recursive: true);
+                        }
                     }
                     catch (Exception directoryDeleteError)
                     {
